Refactor: 테이블명 통일, 코드 포맷팅
Resolves: #12
