+++ LectureFlow/src/main/java/github/nbcamp/lectureflow/domain/lecture/service/ExcelUpload.java
@@ -1,7 +1,6 @@
 package github.nbcamp.lectureflow.domain.lecture.service;
 
 import github.nbcamp.lectureflow.domain.lecture.dto.LectureRequestDto;
-import github.nbcamp.lectureflow.global.entity.Lecture;
 import github.nbcamp.lectureflow.global.enums.Day;
 import github.nbcamp.lectureflow.global.enums.Department;
 import github.nbcamp.lectureflow.global.enums.MajorOrGeneral;
@@ -18,9 +17,9 @@ import java.util.Iterator;
 import java.util.List;
 
 public class ExcelUpload {
-    public static List<LectureRequestDto> exelToLecture(InputStream inputStream){
+    public static List<LectureRequestDto> exelToLecture(InputStream inputStream) {
 
-        try{
+        try {
             // 엑셀 파일 읽어서 Workbook 객체로 만듦
             Workbook workbook = new XSSFWorkbook(inputStream);
             // 첫 번째 시트를 가져옴(시트가 하나라고 가정하고 진행할 예정이기 때문에 인덱스 0 고정)
@@ -31,11 +30,11 @@ public class ExcelUpload {
             //결과 담을 리스트
             List<LectureRequestDto> lectureList = new ArrayList<>();
 
-            //다음 열이 존재하지 않으면 멈춤
-            while(rowIterator.hasNext()){
+            //다음 행이 존재하지 않으면 멈춤
+            while (rowIterator.hasNext()) {
                 // 행 하나를 꺼내서 저장
                 Row nowRow = rowIterator.next();
-                if(rowNumber == 0){
+                if (rowNumber == 0) {
                     rowNumber++;
                     continue;
                 }
@@ -46,10 +45,10 @@ public class ExcelUpload {
                 LectureRequestDto.LectureRequestDtoBuilder lectureRequestDtoBuilder = LectureRequestDto.builder();
 
                 // 다음 셀이 존재하지 않으면 멈춤
-                while(cellIterator.hasNext()){
+                while (cellIterator.hasNext()) {
                     // 셀 하나를 꺼내서 저장
                     Cell nowCell = cellIterator.next();
-                    switch(cellIndex){
+                    switch (cellIndex) {
                         case 0:
                             String uploadMajorOrGeneral = nowCell.getStringCellValue();
                             lectureRequestDtoBuilder.majorOrGeneral(MajorOrGeneral.valueOf(uploadMajorOrGeneral));
@@ -60,7 +59,7 @@ public class ExcelUpload {
                             break;
                         case 2:
                             double gradeLevel = nowCell.getNumericCellValue();
-                            lectureRequestDtoBuilder.gradeLevel((int)gradeLevel);
+                            lectureRequestDtoBuilder.gradeLevel((int) gradeLevel);
                             break;
                         case 3:
                             boolean isForeignLanguage = nowCell.getBooleanCellValue();
@@ -72,7 +71,7 @@ public class ExcelUpload {
                             break;
                         case 5:
                             double grade = nowCell.getNumericCellValue();
-                            lectureRequestDtoBuilder.grade((int)grade);
+                            lectureRequestDtoBuilder.grade((int) grade);
                             break;
                         case 6:
                             String professor = nowCell.getStringCellValue();
@@ -83,20 +82,20 @@ public class ExcelUpload {
                             lectureRequestDtoBuilder.day(Day.valueOf(day));
                             break;
                         case 8:
-                             String startTime = nowCell.getStringCellValue();
+                            String startTime = nowCell.getStringCellValue();
                             lectureRequestDtoBuilder.startTime(LocalTime.parse(startTime));
                             break;
                         case 9:
-                             String endTime = nowCell.getStringCellValue();
+                            String endTime = nowCell.getStringCellValue();
                             lectureRequestDtoBuilder.endTime(LocalTime.parse(endTime));
                             break;
                         case 10:
                             double classroom = nowCell.getNumericCellValue();
-                            lectureRequestDtoBuilder.classroom((int)classroom);
+                            lectureRequestDtoBuilder.classroom((int) classroom);
                             break;
                         case 11:
                             double maxStudent = nowCell.getNumericCellValue();
-                            lectureRequestDtoBuilder.maxStudent((int)maxStudent);
+                            lectureRequestDtoBuilder.maxStudent((int) maxStudent);
                             break;
                     }
                     cellIndex++;
@@ -104,7 +103,7 @@ public class ExcelUpload {
                 lectureList.add(lectureRequestDtoBuilder.build());
             }
             return lectureList;
-        }catch(IOException e){
+        } catch (IOException e) {
             throw new RuntimeException("예외처리 예정이용");
         }
     }
