+++ LectureFlow/src/main/java/github/nbcamp/lectureflow/domain/lecture/service/LectureService.java
@@ -7,7 +7,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
@@ -35,7 +34,7 @@ public class LectureService {
                 lectureRequestDto.getEndTime(),
                 lectureRequestDto.getClassroom(),
                 lectureRequestDto.getMaxStudent()
-                )).toList();
+        )).toList();
         lectureRepository.saveAll(lectureList);
     }
 }
